# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * LuaBox API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface OverviewStats {
  totalScripts: number;
  totalPanels: number;
  totalKeys: number;
  activeUsers: number;
  connectedServers: number;
}

export interface GuildOption {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
}

export interface ChannelOption {
  id: string;
  name: string;
}

export interface BotInviteResponse {
  url: string;
}

export interface SyncResult {
  synced: number;
  total: number;
}

export type ScriptStatus = typeof ScriptStatus[keyof typeof ScriptStatus];


export const ScriptStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export interface Script {
  id: number;
  ownerId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  version: string;
  status: ScriptStatus;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  obfuscatedContent?: string | null;
  createdAt: string;
}

export interface ScriptInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  version: string;
  content?: string;
}

export type ScriptUpdateStatus = typeof ScriptUpdateStatus[keyof typeof ScriptUpdateStatus];


export const ScriptUpdateStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export interface ScriptUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minLength 1 */
  version?: string;
  status?: ScriptUpdateStatus;
  content?: string;
}

export interface Panel {
  id: number;
  ownerId: number;
  scriptId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  discordServerId?: string | null;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  messageId?: string | null;
  requiredRoles?: string[];
  createdAt: string;
}

export interface PanelInput {
  scriptId: number;
  /** @minLength 1 */
  name: string;
  description?: string;
  discordServerId?: string;
  channelId?: string;
  requiredRoles?: string[];
}

export interface PanelUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  discordServerId?: string;
  channelId?: string;
  requiredRoles?: string[];
}

export interface PanelSendInput {
  channelId: string;
}

export type LicenseKeyStatus = typeof LicenseKeyStatus[keyof typeof LicenseKeyStatus];


export const LicenseKeyStatus = {
  active: 'active',
  revoked: 'revoked',
  expired: 'expired',
} as const;

export interface LicenseKey {
  id: number;
  key: string;
  scriptId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  status: LicenseKeyStatus;
  whitelisted?: boolean;
  createdAt: string;
}

export interface KeyInput {
  scriptId: number;
  userId?: number;
  expiresAt?: string;
  whitelisted?: boolean;
}

export interface DiscordServer {
  id: number;
  guildId: string;
  name: string;
  ownerId?: number;
  createdAt: string;
}

export interface WhitelistEntry {
  id: number;
  scriptId: number;
  discordUserId: string;
  /** @nullable */
  addedBy?: string | null;
  createdAt: string;
}

export interface WhitelistInput {
  discordUserId: string;
  addedBy?: string;
}