# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LuaBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /guilds/{guildId}/channels:
    get:
      operationId: listGuildChannels
      tags: [guilds]
      summary: List text channels in a Discord guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of text channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelOption"
...
  /keys:
    get:
      operationId: listKeys
      tags: [keys]
      summary: List all license keys for current user's scripts
      responses:
        "200":
          description: List of license keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LicenseKey"
    post:
      operationId: generateKey
      tags: [keys]
      summary: Generate a new license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyInput"
      responses:
        "201":
          description: Generated license key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"
  /keys/{id}:
    get:
      operationId: getKey
      tags: [keys]
      summary: Get a license key by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"
    delete:
      operationId: deleteKey
      tags: [keys]
      summary: Delete a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /keys/{id}/revoke:
    post:
      operationId: revokeKey
      tags: [keys]
      summary: Revoke a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    User:
      type: object
      required: [id, discordId, username, createdAt]
      properties:
        id:
          type: integer
        discordId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ChannelOption:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...
    LicenseKey:
      type: object
      required: [id, key, scriptId, status, createdAt]
      properties:
        id:
          type: integer
        key:
          type: string
        scriptId:
          type: integer
        userId:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, revoked, expired]
        whitelisted:
          type: boolean
        createdAt:
          type: string
    KeyInput:
      type: object
      required: [scriptId]
      properties:
        scriptId:
          type: integer
        userId:
          type: integer
        expiresAt:
          type: string
        whitelisted:
          type: boolean
...